backend/user_service/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6

sqlalchemy==2.0.35
psycopg2-binary==2.9.9

pydantic==2.9.2
pydantic-settings==2.4.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # thêm dòng này

starlette==0.38.4
python-dotenv==1.0.1
email-validator
